optimize some code
